Moved all DP functionality to a separate DP policy which inherits from policy, separating the knowledge of the MDP DP needs from the other policy algorithms. Added back terminal states to the policy since y'all were right and they do have a right to be there (removing them in the main Policy class requires more knowledge about the MDP than the class should have).
